% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/core.R
\name{addData_TFActivity}
\alias{addData_TFActivity}
\title{Add TF activity data to GRN object using a simplified procedure for estimating it. EXPERIMENTAL.}
\usage{
addData_TFActivity(
  GRN,
  normalization = "cyclicLoess",
  name = "TF_activity",
  forceRerun = FALSE
)
}
\arguments{
\item{GRN}{Object of class \code{\linkS4class{GRN}}}

\item{normalization}{Character. Default \code{cyclicLoess}. One of \code{cyclicLoess}, \code{sizeFactors}, \code{quantile}, or \code{none}. Normalization procedure.}

\item{name}{Name in object under which it should be stored. This corresponds to the \code{connectionType} afterwards that some functions iterate over.}

\item{forceRerun}{\code{TRUE} or \code{FALSE}. Default \code{FALSE}. Force execution, even if the GRN object already contains the result. Overwrites the old results.}
}
\value{
The same \code{\linkS4class{GRN}} object, with added data from this function.
}
\description{
Add TF activity data to GRN object using a simplified procedure for estimating it. EXPERIMENTAL.
}
