% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/core.R
\name{getParameters}
\alias{getParameters}
\title{Retrieve parameters for previously used function calls and general parameters for a \code{\linkS4class{GRN}} object.}
\usage{
getParameters(GRN, type = "function", name = NULL)
}
\arguments{
\item{GRN}{Object of class \code{\linkS4class{GRN}}}

\item{type}{Character. Default \code{function}. Either \code{function}, \code{parameter}, or \code{all}. When set to \code{function}, a valid \code{GRaNIE} function name must be given that has been run before. \code{parameter} indicates a particular parameter name is returned (as specified in \code{GRN@config})), while \code{all} returns all parameters.}

\item{name}{Character. Name of parameter or function name to retrieve. Ignored if \code{type} == \code{all}.}
}
\value{
The same \code{\linkS4class{GRN}} object, with added data from this function.
}
\description{
Retrieve parameters for previously used function calls and general parameters for a \code{\linkS4class{GRN}} object.
}
\examples{
GRN = loadExampleObject()
getParameters(GRN, type = "function")

}
