% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helperFunctions.R
\name{installSuggestedPackages}
\alias{installSuggestedPackages}
\title{Automatically install all suggested packages for the \code{GRaNIE} package}
\usage{
installSuggestedPackages(includeDeveloperPackages = FALSE)
}
\arguments{
\item{includeDeveloperPackages}{\code{TRUE} or \code{FALSE}. Default \code{FALSE}. Should suggested developer packages (\code{BiocStyle}, \code{knitr}, \code{testthat}) also be installed?}
}
\value{
Nothing, but all packages should be available upon execution of the function
}
\description{
This helper function attempts to install all packages listed under \code{Suggests}. 
\strong{Note that installing all these packages may take a considerable amount of time, so please be patient.}
\strong{Also, only packages are attempted to be installed when they are not yet installed.}
\strong{If the automated installation doesnt work, you have to manually install missing packages.}
The list of packages contains a list of packages that are not strictly necessary for the workflow of \code{GRaNIE}, 
but which enhance the functionality, may be required depending on certain parameters (such as your genome assembly version), 
may be required only when using a particular functionality (such as the \code{robust} package for robust regression),
or speed-up the computation (via \code{BiocParallel}).
}
\examples{
\dontrun{
installSuggestedPackages()
}
}
