% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/core.R
\name{loadExampleObject}
\alias{loadExampleObject}
\title{Load example GRN dataset}
\usage{
loadExampleObject()
}
\value{
An example \code{\linkS4class{GRN}} object
}
\description{
Load example GRN dataset
}
\examples{
GRN = loadExampleObject()
}
