% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DataClasses.R
\name{nTFs}
\alias{nTFs}
\alias{TFs}
\title{Get the number of TFs for a \code{\linkS4class{GRN}} object.}
\usage{
nTFs(GRN)
}
\arguments{
\item{GRN}{Object of class \code{\linkS4class{GRN}}}
}
\value{
Integer. Number of TFs that are defined in the \code{\linkS4class{GRN}} object.
}
\description{
Returns the number of TFs  from the provided TFBS data in the \code{\linkS4class{GRN}} object.
}
\examples{
# See the Workflow vignette on the GRaNIE website for examples
GRN = loadExampleObject()
nTFs(GRN)
}
\seealso{
\code{\link{nGenes}}

\code{\link{nPeaks}}
}
