% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/core.R
\name{overlapPeaksAndTFBS}
\alias{overlapPeaksAndTFBS}
\title{Overlap peaks and TFBS for a \code{\linkS4class{GRN}} object}
\usage{
overlapPeaksAndTFBS(GRN, nCores = 2, forceRerun = FALSE)
}
\arguments{
\item{GRN}{Object of class \code{\linkS4class{GRN}}}

\item{nCores}{Integer >0. Default 1. Number of cores to use.}

\item{forceRerun}{\code{TRUE} or \code{FALSE}. Default \code{FALSE}. Force execution, even if the GRN object already contains the result. Overwrites the old results.}
}
\value{
The same \code{\linkS4class{GRN}} object, with added data from this function.
}
\description{
Overlap peaks and TFBS for a \code{\linkS4class{GRN}} object
}
\examples{
# See the Workflow vignette on the GRaNIE website for examples
# GRN = loadExampleObject()
# GRN = overlapPeaksAndTFBS(GRN, nCores = 2, forceRerun = FALSE)
}
