% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot.R
\name{visualizeGRN}
\alias{visualizeGRN}
\title{Visualize a filtered GRN (in development)}
\usage{
visualizeGRN(
  GRN,
  permuted,
  file,
  title = NULL,
  maxRowsToPlot = 500,
  useDefaultMetadata = TRUE,
  vertice_color_TFs = NULL,
  vertice_color_peaks = NULL,
  vertice_color_genes = NULL,
  plotAsPDF = TRUE,
  pdf_width = 12,
  pdf_height = 12,
  forceRerun = FALSE
)
}
\arguments{
\item{GRN}{Object of class \code{\linkS4class{GRN}}}

\item{permuted}{\code{TRUE} or \code{FALSE}. Default \code{FALSE}. Should the permuted data be taken (\code{TRUE}) or the non-permuted, original one (\code{FALSE})?}

\item{file}{TODO}

\item{title}{TODO}

\item{maxRowsToPlot}{TODO}

\item{useDefaultMetadata}{TODO}

\item{vertice_color_TFs}{TODO}

\item{vertice_color_peaks}{TODO}

\item{vertice_color_genes}{TODO}

\item{plotAsPDF}{\code{TRUE} or \code{FALSE}. Default \code{TRUE}.Should the plots be printed to a PDF file? If set to \code{TRUE}, a PDF file is generated, the name of which depends on the value of \code{basenameOutput}. If set to \code{FALSE}, all plots are printed to the currently active device. Note that most functions print more than one plot, which means you may only see the last plot depending on your active graphics device.}

\item{pdf_width}{Number. Default 12. Width of the PDF, in cm.}

\item{pdf_height}{Number. Default 12. Height of the PDF, in cm.}

\item{forceRerun}{\code{TRUE} or \code{FALSE}. Default \code{FALSE}. Force execution, even if the GRN object already contains the result. Overwrites the old results.}
}
\value{
TODO
}
\description{
Visualize a filtered GRN (in development)
}
