% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/core.R
\name{addSNPData}
\alias{addSNPData}
\title{Add SNP information to a \code{\linkS4class{GRN}} object.}
\usage{
addSNPData(GRN, SNP_IDs, EnsemblVersion = NULL, forceRerun = FALSE)
}
\arguments{
\item{GRN}{Object of class \code{\linkS4class{GRN}}}

\item{SNP_IDs}{Character vector. No default. Vector of SNP IDs (rsID) that should be integrated and overlapped with the peaks.}

\item{EnsemblVersion}{\code{NULL} or Character(1). Default \code{NULL}. Only relevant if \code{source} is not set to \code{custom}, ignored otherwise. The Ensembl version to use for the retrieval of gene IDs from their provided database names (e.g., JASPAR) via \code{biomaRt}.
By default (\code{NULL}), the newest version is selected for the most recent genome assembly versions is used (see \code{biomaRt::listEnsemblArchives()} for supported versions). This parameter can override this to use a custom (older) version instead.}

\item{forceRerun}{\code{TRUE} or \code{FALSE}. Default \code{FALSE}. Force execution, even if the GRN object already contains the result. Overwrites the old results.}
}
\value{
An updated \code{\linkS4class{GRN}} object, with additional information added from this function.
}
\description{
This function accepts a vector of SNP IDs (rsID), retrieves their genomic positions and 
overlaps them with the peaks to extend the peak metadata (`GRN@data$peaks$counts_metadata`) by storing the number, positions and rsids of all
overlapping SNPs per peak  (new columns starting with `SNP_`).
}
\details{
`biomaRt` is used to retrieve genomic positions for the user-defined SNPs, which can take a long time depending
on the number of SNPs provided.
}
\examples{
# See the Workflow vignette on the GRaNIE website for examples
GRN = loadExampleObject()
GRN = addSNPData(GRN, SNP_IDs = c("rs7570219", "rs6445264", "rs12067275"), forceRerun = FALSE)

}
