% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/core.R
\name{addTFBS}
\alias{addTFBS}
\title{Add TFBS to a \code{\linkS4class{GRN}} object}
\usage{
addTFBS(
  GRN,
  motifFolder,
  TFs = "all",
  nTFMax = NULL,
  filesTFBSPattern = "_TFBS",
  fileEnding = ".bed",
  forceRerun = FALSE
)
}
\arguments{
\item{GRN}{Object of class \code{\linkS4class{GRN}}}

\item{motifFolder}{Character. No default. Path to the folder that contains the TFBS predictions. The files must be in BED format, 6 columns, one file per TF. See the other parameters for more details.}

\item{TFs}{Character vector. Default \code{all}. Vector of TF names to include. The special keyword \code{all} can be used to include all TF found in the folder as specified by \code{motifFolder}. If \code{all} is specified anywhere, all TFs will be included. TF names must otherwise match the file names that are found in the folder, without the file suffix.}

\item{nTFMax}{\code{NULL} or integer. Default \code{NULL}. Maximal number of TFs to import. Can be used for testing purposes, e.g., setting to 5 only imports 5 TFs even though the whole \code{motifFolder} has many more TFs defined.}

\item{filesTFBSPattern}{Character. Default \code{"_TFBS"}. Suffix for the file names in the TFBS folder that is not part of the TF name. Can be empty. For example, for the TF CTCF, if the file is called \code{CTCF.all.TFBS.bed}, set this parameter to \code{".all.TFBS"}.}

\item{fileEnding}{Character. Default \code{".bed"}. File ending for the files from the motif folder.}

\item{forceRerun}{\code{TRUE} or \code{FALSE}. Default \code{FALSE}. Force execution, even if the GRN object already contains the result. Overwrites the old results.}
}
\value{
The same \code{\linkS4class{GRN}} object, with added data from this function.
}
\description{
Add TFBS to a \code{\linkS4class{GRN}} object
}
\examples{
# See the Workflow vignette on the GRaNIE website for examples
}
