% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/core.R
\name{changeOutputDirectory}
\alias{changeOutputDirectory}
\title{Change the output directory of a GRN object}
\usage{
changeOutputDirectory(GRN, outputDirectory = ".")
}
\arguments{
\item{GRN}{Object of class \code{\linkS4class{GRN}}}

\item{outputDirectory}{Character. Default \code{.}. New output directory for all output files unless overwritten via the parameter \code{outputFolder}.}
}
\value{
The same \code{\linkS4class{GRN}} object, with the output directory being adjusted accordingly
}
\description{
Change the output directory of a GRN object
}
\examples{
GRN = loadExampleObject()
GRN = changeOutputDirectory(GRN, outputDirectory = ".")
}
