% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/core.R
\name{deleteIntermediateData}
\alias{deleteIntermediateData}
\title{Optional convenience function to delete intermediate data from the function \link{AR_classification_wrapper} and summary statistics that may occupy a lot of space}
\usage{
deleteIntermediateData(GRN)
}
\arguments{
\item{GRN}{Object of class \code{\linkS4class{GRN}}}
}
\value{
TODO
}
\description{
Optional convenience function to delete intermediate data from the function \link{AR_classification_wrapper} and summary statistics that may occupy a lot of space
}
