% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/core.R
\name{getGRNConnections}
\alias{getGRNConnections}
\title{Extract connections from a \code{\linkS4class{GRN}} object}
\usage{
getGRNConnections(
  GRN,
  type = "all.filtered",
  permuted = FALSE,
  include_TF_gene_correlations = FALSE
)
}
\arguments{
\item{GRN}{Object of class \code{\linkS4class{GRN}}}

\item{type}{Character. Default "all.filtered". Must be one of "TF_peaks", "peak_genes", "all.filtered". The type of connections to retrieve.}

\item{permuted}{\code{TRUE} or \code{FALSE}. Default \code{FALSE}. Should the permuted data be taken (\code{TRUE}) or the non-permuted, original one (\code{FALSE})?}

\item{include_TF_gene_correlations}{Logical. \code{TRUE} or \code{FALSE}.  Should TFs and gene correlations be returned as well? If set to \code{TRUE}, they must have been computed beforehand with \code{\link{add_TF_gene_correlation}}.}
}
\value{
A data frame with the connections. Importantly, this function does NOT return a \code{\linkS4class{GRN}} object.
}
\description{
Extract connections from a \code{\linkS4class{GRN}} object
}
\examples{
GRN = loadExampleObject()
GRN_con.all = getGRNConnections(GRN, include_TF_gene_correlations = TRUE)
}
