% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/network.R
\name{getTopNodes}
\alias{getTopNodes}
\title{Retrieve top Nodes in the filtered \code{\linkS4class{GRN}}}
\usage{
getTopNodes(GRN, nodeType, rankType, n = 0.1, use_TF_gene_network = TRUE)
}
\arguments{
\item{GRN}{Object of class \code{\linkS4class{GRN}}}

\item{nodeType}{Character. One of: "gene", "TF".}

\item{rankType}{Character. One of: "degree", "EV". This parameter will determine the criterion to be used to identify the "top" nodes. If set to "degree", the function will select top nodes based on the number of connections they have, i.e. based on their degree-centrality. If set to "EV" it will select the top nodes based on their eigenvector-centrality score in the network.}

\item{n}{Numeric. Default 0.1. If this parameter is passed as a value between (0,1), it is treated as a percentage of top nodes. If the value is passed as an integer it will be treated as the number of top nodes.}

\item{use_TF_gene_network}{\code{TRUE} or \code{FALSE}. Default \code{TRUE}. Should the TF-gene network be used (\code{TRUE}) or the TF-peak-gene network (\code{FALSE})?}
}
\value{
A dataframe with the node names and the corresponding scores used to rank them
}
\description{
Retrieve top Nodes in the filtered \code{\linkS4class{GRN}}
}
\examples{
GRN = loadExampleObject()
topGenes = getTopNodes(GRN, nodeType = "gene", rankType = "degree", n = 3)
topTFs = getTopNodes(GRN, nodeType = "TF", rankType = "EV", n = 5)
}
