% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/core.R
\name{importTFData}
\alias{importTFData}
\title{Import externally derived TF Activity data. EXPERIMENTAL.}
\usage{
importTFData(
  GRN,
  data,
  name,
  idColumn = "ENSEMBL",
  nameColumn = "TF.name",
  normalization = "none",
  forceRerun = FALSE
)
}
\arguments{
\item{GRN}{Object of class \code{\linkS4class{GRN}}}

\item{data}{Data frame. No default. Data with TF data.}

\item{name}{Name in object under which it should be stored. This corresponds to the \code{connectionType} afterwards that some functions iterate over.}

\item{idColumn}{Character. Default \code{ENSEMBL}. Name of the ID column. Must not be unique as some TFs may correspond to the same ID.}

\item{nameColumn}{Character. Default \code{TF.name}. Must be unique for each TF / row.}

\item{normalization}{Character. Default \code{cyclicLoess}. One of \code{cyclicLoess}, \code{sizeFactors}, \code{quantile}, or \code{none}. 
Normalization procedure. When set to \code{cyclicLoess}, the \code{csaw} package is required (as it is listed under \code{Suggests}, it may not be installed).}

\item{forceRerun}{\code{TRUE} or \code{FALSE}. Default \code{FALSE}. Force execution, even if the GRN object already contains the result. Overwrites the old results.}
}
\value{
An updated \code{\linkS4class{GRN}} object, with added data from this function.
}
\description{
We do not yet provide full support for this function. It is currently being tested. Use at our own risk.
}
